requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
cffi==1.12.1
chardet==3.0.4
discord.py==1.0.0a1691+gf686924
humanize==0.5.1
idna==2.8
import-expression==0.3.8.post2
jishaku==1.6.1
multidict==4.5.2
pycparser==2.19
PyNaCl==1.2.1
six==1.12.0
websockets==6.0
yarl==1.3.0
